Fix : userinfo, Geport_view, iGeport_view 수정
