+++ src/app/[id]/page.js
