+++ src/app/view/iGeport_view/page.js
@@ -4,13 +4,13 @@ import {useEffect, useState} from "react";
 import {Fira_Sans} from "next/dist/compiled/@next/font/dist/google";
 import Image from "next/image";
 
-export default function Geport_view(){
+export default function iGeport_view(){
     const router = useRouter();
-    const [loading, setLoading] = useState(true);
+    const [loading, setLoading] = useState(true);//3초 로딩 해보자
     const [paging, setPaging] = useState(0);
     const [isInfo, setIsInfo] = useState(true);
     const [answer, setAnswer] = useState([
-        {id : 1, content: ""}, {id : 2, content: ""}, {id : 3, content: ""}, {id : 4, content: ""}, {id : 5, content: ""}
+        {id : 1, content: ""}, {id : 2, content: ""}, {id : 3, content: ""}, {id : 4, content: ""}, {id : 5, content: ""}, {id : 6, content: ""}
     ]);
 
     useEffect(() => {
@@ -24,7 +24,7 @@ export default function Geport_view(){
     }, []); // This effect runs only once when the component mounts
 
     useEffect(() => {
-        const isPage = paging >= 1 && paging <= 5;
+        const isPage = paging >= 1 && paging <= 6;
         setIsInfo(!isPage);
     }, [paging]);
 
@@ -34,6 +34,7 @@ export default function Geport_view(){
         );
         setAnswer(updatedAnswers);
     };
+
 
     const pageComponent = (page) => {
         switch (page){
@@ -57,6 +58,10 @@ export default function Geport_view(){
                 return (
                     <Fifth answer={answer} page={page} update={update}/>
                 )
+            case 6:
+                return (
+                    <Six answer={answer} page={page} update={update}/>
+                )
             default:
                 return(
                     <Information/>
@@ -72,7 +77,7 @@ export default function Geport_view(){
             flexDirection: 'column',
             weight: "100%"
         }}>
-            {loading ? (
+            {loading ? ( // Display loading screen for 3 seconds
                 <LoadingScreen />
             ) : (
                 <>
@@ -81,7 +86,7 @@ export default function Geport_view(){
                             <button style={{ fontSize: "1.8em" }} onClick={() => {
                                 if (paging < 1) {
                                     router.refresh();
-                                    router.push('/hidden-code');
+                                    router.push('/userinfo');
                                 } else {
                                     setPaging(prev => prev - 1)
                                 }
@@ -96,7 +101,7 @@ export default function Geport_view(){
                     {pageComponent(paging)}
                     {isInfo ? <div style={{ display: 'flex', justifyContent: 'center', gap: '8px', margin: '25px 0' }}></div> :
                         <div style={{ display: 'flex', justifyContent: 'center', gap: '8px', margin: '20px 0' }}>
-                            {Array.from({ length: 5 }, (_, index) => (
+                            {Array.from({ length: 6 }, (_, index) => (
                                 <span
                                     key={index}
                                     style={{
@@ -125,21 +130,22 @@ export default function Geport_view(){
                         opacity: 1
                     }}
                          onClick={() => {
-                             if (paging > 4) {
+                             if (paging > 5) {
                                  router.refresh();
-                                 router.push('/hidden-code');
+                                 router.push('/userinfo'); // 다음 페이지로 이동하자
                              }
                              else {
                                  setPaging(prev => prev + 1)
                              }
                          }}>
-                        {paging === 5 ? "Geport 다운받기" : "다음으로"}
+                        {paging === 6 ? "iGeport 다운받기" : "다음으로"}
                     </div>
                 </>
             )}
         </div>
     )
 }
+
 
 export function Information() {
     return (
@@ -148,7 +154,7 @@ export function Information() {
         }}>
             <div style={{width: "100%", height: "15%", fontSize: "1.6em",
                 marginBottom: "10%" , paddingLeft: "10%", paddingRight: "10%"}}>
-                (소개글 한 문장), Username 님의 Geport
+                (소개글 한문장), Uswername 님의 iGeport
             </div>
             <div style={{
                 width: "100%",
@@ -160,8 +166,7 @@ export function Information() {
                 paddingRight: "10%",
                 marginTop:"5%"
             }}>
-                Username 님의 블로글와 답변을 바탕으로
-                퍼스널 브랜딩에 필요한 분석 솔루션을 준비했어요
+                Username 님의 블로그와 답변을 바탕으로 퍼스널 브랜딩에 필요한 심리 보고서를 준비했어요.
             </div>
         </div>
     )
@@ -181,8 +186,11 @@ export function First({answer, page, update}){
             }// 스크롤을 허용하는 부분
         }}>
             <div style={{width: "100%", height: "10%", fontSize: "1.5em",
-                marginBottom: "10%" , paddingLeft: "10%", paddingRight: "10%"}}>
-                Username 님은 이런 사람이 되고 싶어해요
+                marginBottom: "20px" , paddingLeft: "10%", paddingRight: "10%"}}>
+                Username 님의 블로그를 분석했어요
+            </div>
+            <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <Image src={"/image/거북이랑 수영을.png"} width={400} height={100} alt={"logo"}/>
             </div>
             <div style={{
                 width: "100%",
@@ -194,7 +202,52 @@ export function First({answer, page, update}){
                 paddingRight: "10%",
                 marginTop:"5%"
             }}>
-                Lorem ipsum dolor sit amet consectetur. Amet pharetra consequat diam nunc eget accumsan fermentum enim quam. Convallis scelerisque pellentesque mi commodo in. Nulla nunc cursus ullamcorper amet aliquam diam turpis tempus nunc. Faucibus venenatis neque morbi amet leo diam.
+                그저 예시문
+            </div>
+            <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <Image src={"/image/gc_sky.jpg"} width={400} height={100} alt={"logo"}/>
+            </div>
+            <div style={{
+                width: "100%",
+                height: "10%",
+                fontSize: "1em",
+                marginBottom: "5%",
+                color: "#C6C6C6",
+                paddingLeft: "10%",
+                paddingRight: "10%",
+                marginTop:"5%"
+            }}>
+                그저 예시문
+            </div>
+            <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <Image src={"/image/거북이랑 수영을.png"} width={400} height={100} alt={"logo"}/>
+            </div>
+            <div style={{
+                width: "100%",
+                height: "10%",
+                fontSize: "1em",
+                marginBottom: "5%",
+                color: "#C6C6C6",
+                paddingLeft: "10%",
+                paddingRight: "10%",
+                marginTop:"5%"
+            }}>
+                그저 예시문
+            </div>
+            <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <Image src={"/image/gc_sky.jpg"} width={400} height={100} alt={"logo"}/>
+            </div>
+            <div style={{
+                width: "100%",
+                height: "10%",
+                fontSize: "1em",
+                marginBottom: "5%",
+                color: "#C6C6C6",
+                paddingLeft: "10%",
+                paddingRight: "10%",
+                marginTop:"5%"
+            }}>
+                그저 예시문
             </div>
         </div>
     )
@@ -215,8 +268,11 @@ export function Second({answer, page, update}) {
         }}>
             <div style={{width: "100%", height: "10%", fontSize: "1.5em",
                 marginBottom: "10%" , paddingLeft: "10%", paddingRight: "10%"}}>
-                Username 님의 좌우명을
-                Geport가 분석했어요
+                이 기간 동안 username 님의
+                감정곡선을 보여드려요
+            </div>
+            <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <Image src={"/image/거북이랑 수영을.png"} width={400} height={100} alt={"logo"}/>
             </div>
             <div style={{
                 width: "100%",
@@ -228,7 +284,8 @@ export function Second({answer, page, update}) {
                 paddingRight: "10%",
                 marginTop:"5%"
             }}>
-                Lorem ipsum dolor sit amet consectetur. Amet pharetra consequat diam nunc eget accumsan fermentum enim quam. Convallis scelerisque pellentesque mi commodo in. Nulla nunc cursus ullamcorper amet aliquam diam turpis tempus nunc. Faucibus venenatis neque morbi amet leo diam.
+                username 님은 여행 일자에 따라 행복, 만족감, 기쁨의 지수가 점차 증가합니다.
+                반면 우울감 지수는 낮아지며, 이는 여행이 긍정적인 경험 이었음을 시사합니다.
             </div>
         </div>
     )
@@ -252,6 +309,9 @@ export function Third({answer, page, update}) {
                 이 기간 동안 username 님의
                 감정 SOS를 알려드려요
             </div>
+            <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <Image src={"/image/거북이랑 수영을.png"} width={400} height={100} alt={"logo"}/>
+            </div>
             <div style={{
                 width: "100%",
                 height: "10%",
@@ -262,7 +322,9 @@ export function Third({answer, page, update}) {
                 paddingRight: "10%",
                 marginTop:"5%"
             }}>
-                Lorem ipsum dolor sit amet consectetur. Amet pharetra consequat diam nunc eget accumsan fermentum enim quam. Convallis scelerisque pellentesque mi commodo in. Nulla nunc cursus ullamcorper amet aliquam diam turpis tempus nunc. Faucibus venenatis neque morbi amet leo diam.</div>
+                이 그래프는 username 님의 부정적 감정 상태를 나타냅니다.
+                스트레스, 불안, 슬픔, 피로와 같은 감정들이 정상 범위 기준선인 200점 아래이며, 이는 여행이 username 님께 긍정적인 영향을 미쳤음을 시사합니다.
+            </div>
         </div>
     )
 }
@@ -272,11 +334,18 @@ export function Fourth({answer, page, update}) {
 
     return (
         <div style={{
-            height: '100vh'
+            height: '100vh',
+            overflowY: 'auto', // 수직 스크롤바만 표시되도록 설정
+            scrollbarWidth: 'none', // Firefox에서 스크롤바를 숨기기 위한 설정
+            '-ms-overflow-style': 'none', // IE에서 스크롤바를 숨기기 위한 설정
+            '&::-webkit-scrollbar': { // Webkit 엔진 기반 브라우저에서 스크롤바를 숨기기 위한 설정
+                display: 'none'
+            }// 스크롤을 허용하는 부분
         }}>
             <div style={{width: "100%", height: "10%", fontSize: "1.5em",
                 marginBottom: "10%" , paddingLeft: "10%", paddingRight: "10%"}}>
-                Username 님의 삶을 그래프로 표현했어요
+                블로그를 작성할 당시
+                username 님은 이런 모습이었어요.
             </div>
             <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
                 <Image src={"/image/거북이랑 수영을.png"} width={400} height={100} alt={"logo"}/>
@@ -291,8 +360,8 @@ export function Fourth({answer, page, update}) {
                 paddingRight: "10%",
                 marginTop:"5%"
             }}>
-                Lorem ipsum dolor sit amet consectetur. Amet pharetra consequat diam nunc eget accumsan fermentum enim quam.
-            </div>
+                Big 5 점수 설명
+                </div>
         </div>
     )
 }
@@ -312,7 +381,11 @@ export function Fifth({answer, page, update}){
         }}>
             <div style={{width: "100%", height: "10%", fontSize: "1.5em",
                 marginBottom: "10%" , paddingLeft: "10%", paddingRight: "10%"}}>
-                Username 님의 Geport 퍼스널 브랜딩 솔루션은
+                이 기간 동안 username 님을
+                행복하게 만든 것을 모아봤어요
+            </div>
+            <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <Image src={"/image/거북이랑 수영을.png"} width={400} height={100} alt={"logo"}/>
             </div>
             <div style={{
                 width: "100%",
@@ -324,23 +397,57 @@ export function Fifth({answer, page, update}){
                 paddingRight: "10%",
                 marginTop:"5%"
             }}>
-                Lorem ipsum dolor sit amet consectetur. Amet pharetra consequat diam nunc eget accumsan fermentum enim quam. Convallis scelerisque pellentesque mi commodo in. Nulla nunc cursus ullamcorper amet aliquam diam turpis tempus nunc. Faucibus venenatis neque morbi amet leo diam.
+                </div>
+        </div>
+    )
+}
+export function Six({answer, page, update}){
+    const myContent = answer.find(ans => ans.id === page)?.content || "";
+
+    return (
+        <div style={{
+            height: '100vh',
+            overflowY: 'auto', // 수직 스크롤바만 표시되도록 설정
+            scrollbarWidth: 'none', // Firefox에서 스크롤바를 숨기기 위한 설정
+            '-ms-overflow-style': 'none', // IE에서 스크롤바를 숨기기 위한 설정
+            '&::-webkit-scrollbar': { // Webkit 엔진 기반 브라우저에서 스크롤바를 숨기기 위한 설정
+                display: 'none'
+            }// 스크롤을 허용하는 부분
+        }}>
+            <div style={{width: "100%", height: "10%", fontSize: "1.5em",
+                marginBottom: "10%" , paddingLeft: "10%", paddingRight: "10%"}}>
+                username 님의
+                iGeport 감정 분석 솔루션은
+            </div>
+            <div style={{
+                width: "100%",
+                height: "10%",
+                fontSize: "1em",
+                marginBottom: "5%",
+                color: "#C6C6C6",
+                paddingLeft: "10%",
+                paddingRight: "10%",
+                marginTop:"5%"
+            }}>
+                Lorem ipsum dolor sit amet consectetur. Amet pharetra consequat diam numc eget accumsan fermentum enim quam.
+                Convallis scelerisque pellentesque mi commodo in.
+                Nulla nunc cursus ullamcorper amet aliquam diam turpis tempus nunc. Faucibus venenatis neque morbi amet leo diam.
+
             </div>
         </div>
     )
 }
-
 export function LoadingScreen() {
     return (
         <div style={{ backgroundColor: "#181818", height: '100vh', color: 'white', display: 'flex', justifyContent: 'center', alignItems: 'center', flexDirection: 'column' }}>
             <div style={{ marginTop: '30px' }}>
                 <Image src={"/image/logo3.svg"} width={120} height={120} alt="Logo" />
             </div>
-            <h2 style={{ color: 'white', fontSize: '24px', padding: '40px' }}>Geport 결과 보고서</h2>
+            <h2 style={{ color: 'white', fontSize: '24px', padding: '40px' }}>iGeport 결과 보고서</h2>
             <div style={{ marginTop: '30px', textAlign: 'center' }}>
-                <span style={{ color: "#A8A8A8" }}>작성하신 블로그를 바탕으로</span>
+                <span style={{ color: "#A8A8A8" }}>iGeport는 총 여섯 단계로</span>
                 <br />
-                <span style={{ color: "#A8A8A8" }}>퍼스널 브랜딩 솔루션을 만들어 드립니다.</span>
+                <span style={{ color: "#A8A8A8" }}>블로그에 드러난 당신의 심리를 분석합니다.</span>
             </div>
         </div>
     )
