+++ src/app/view/page.js
@@ -1,29 +1,29 @@
-"use client"
-import Image from "next/image";
-import Link from "next/link";
-import {useState} from "react";
-export default function View() {
-    //const [paging, setPaging] = useState(0);
-    return (
-
-         <div style={{backgroundColor: "#181818", height: '100vh', color: 'white', display: 'flex', justifyContent: 'center', alignItems: 'center', flexDirection: 'column'}}>
-             <div style = {{marginTop : '30px'}}>
-                 <Image src={"image/logo3.svg"} width={120} height={120}/>
-             </div>
-             <h2 style={{color:'white', fontSize: '24px', padding : '40px'}}>히든코드를 받으셨나요?</h2>
-             <div style={{marginTop: '30px', textAlign: 'center'}}>
-                 <span style={{color: "#A8A8A8"}}>Geport를 열람할 수 있는 히든코드를</span>
-                 <br />
-                 <span style={{color: "#A8A8A8"}}>입력하신 전화번호로 보내드립니다.</span>
-             </div>
-             <div style={{marginTop : '30px', textAlign: 'center'}}>
-                 <span style={{color: "#A8A8A8"}}>아직 히든코드를 받지 못했다면</span>
-                 <br />
-                 <span style={{color: "#A8A8A8"}}>조금만 더 기다려주세요.</span>
-             </div>
-             <Link href="/hidden-code">
-                 <button style={{color: '#0c0c0c', backgroundColor: '#1AE57C', padding: '10px 30px', borderRadius: '15px', width: '400px', fontWeight: 'bold', marginTop: '150px'}}>히든코드로 Geport 조회하기</button>
-             </Link>
-         </div>
-    );
-}
+// "use client"
+// import Image from "next/image";
+// import Link from "next/link";
+// import {useState} from "react";
+// export default function View() {
+//     //const [paging, setPaging] = useState(0);
+//     return (
+//
+//          <div style={{backgroundColor: "#181818", height: '100vh', color: 'white', display: 'flex', justifyContent: 'center', alignItems: 'center', flexDirection: 'column'}}>
+//              <div style = {{marginTop : '30px'}}>
+//                  <Image src={"image/logo3.svg"} width={120} height={120}/>
+//              </div>
+//              <h2 style={{color:'white', fontSize: '24px', padding : '40px'}}>히든코드를 받으셨나요?</h2>
+//              <div style={{marginTop: '30px', textAlign: 'center'}}>
+//                  <span style={{color: "#A8A8A8"}}>Geport를 열람할 수 있는 히든코드를</span>
+//                  <br />
+//                  <span style={{color: "#A8A8A8"}}>입력하신 전화번호로 보내드립니다.</span>
+//              </div>
+//              <div style={{marginTop : '30px', textAlign: 'center'}}>
+//                  <span style={{color: "#A8A8A8"}}>아직 히든코드를 받지 못했다면</span>
+//                  <br />
+//                  <span style={{color: "#A8A8A8"}}>조금만 더 기다려주세요.</span>
+//              </div>
+//              <Link href="/hidden-code">
+//                  <button style={{color: '#0c0c0c', backgroundColor: '#1AE57C', padding: '10px 30px', borderRadius: '15px', width: '400px', fontWeight: 'bold', marginTop: '150px'}}>히든코드로 Geport 조회하기</button>
+//              </Link>
+//          </div>
+//     );
+// }
